add option to allow shaping on execute_mdx_dataframe_shaped.
execute_mdx_dataframe d can be run with shaped=True to retrieve a similar but not fully equivalent result to execute_mdx_dataframe_shaped.

This allows to retrieve a shaped dataframe while specifying options that are not available in execute_mdx_dataframe_shaped. One example is the usage of use_iterative_json=True which optimizes memory usages during string to json / dict conversion prior to creation of the dataframe.

In the current commit there are some differences in behaviour between execute_mdx_dataframe(shaped=True) and execute_mdx_dataframe_shaped. Those include:
- dtypes are inferred differently
- df.column.name is
equal to the dimension put on the column.

Not tested for multiple dimensions with multiple elements on the columns.